Consolidate test fixtures into root conftest.py
Move all shared fixtures from per-app conftest files into a single
root-level service/conftest.py, keeping only genuinely app-local
conftests (login, integration, nation, variant). Merge the duplicated
game/phase/member/supply_center factories into single definitions and
unify fixture naming: callable-returning fixtures use the *_factory
suffix (replacing make_* and bare names), and the redundant
function-scoped user/client overrides in draw_proposal are dropped in
favor of the session-scoped ones.

Add MD5PasswordHasher to the session-wide test settings override,
cutting the full suite from 6m28s to 1m41s (1494 tests).

Document the fixture layout and naming convention in CLAUDE.md.
