+++ service/game/tests/test_game_delete.py
@@ -11,8 +11,8 @@ delete_viewname = "game-delete"
 class TestGameDeleteView:
 
     @pytest.mark.django_db
-    def test_delete_sandbox_game_as_creator(self, authenticated_client, sandbox_game):
-        game = sandbox_game()
+    def test_delete_sandbox_game_as_creator(self, authenticated_client, sandbox_game_factory):
+        game = sandbox_game_factory()
         url = reverse(delete_viewname, args=[game.id])
         response = authenticated_client.delete(url)
         assert response.status_code == status.HTTP_204_NO_CONTENT
@@ -20,9 +20,9 @@ class TestGameDeleteView:
 
     @pytest.mark.django_db
     def test_delete_sandbox_game_cascades_related_data(
-        self, authenticated_client, sandbox_game
+        self, authenticated_client, sandbox_game_factory
     ):
-        game = sandbox_game()
+        game = sandbox_game_factory()
         phase_ids = list(game.phases.values_list("id", flat=True))
         member_ids = list(game.members.values_list("id", flat=True))
 
@@ -46,9 +46,9 @@ class TestGameDeleteView:
 
     @pytest.mark.django_db
     def test_delete_sandbox_game_as_non_creator(
-        self, authenticated_client_for_secondary_user, sandbox_game
+        self, authenticated_client_for_secondary_user, sandbox_game_factory
     ):
-        game = sandbox_game()
+        game = sandbox_game_factory()
         url = reverse(delete_viewname, args=[game.id])
         response = authenticated_client_for_secondary_user.delete(url)
         assert response.status_code == status.HTTP_403_FORBIDDEN
@@ -56,9 +56,9 @@ class TestGameDeleteView:
 
     @pytest.mark.django_db
     def test_delete_sandbox_game_unauthenticated(
-        self, unauthenticated_client, sandbox_game
+        self, unauthenticated_client, sandbox_game_factory
     ):
-        game = sandbox_game()
+        game = sandbox_game_factory()
         url = reverse(delete_viewname, args=[game.id])
         response = unauthenticated_client.delete(url)
         assert response.status_code == status.HTTP_401_UNAUTHORIZED
